+++ CHANGES.rst
@@ -5,6 +5,8 @@ Unreleased
 ------------
 - Added `bulk_update_with_history` utility function (gh-650)
 - Add default user and default change reason to `bulk_create_with_history` and `bulk_update_with_history`
+- Start using `_change_reason` instead of `changeReason` to add change reasons to historical
+objects. `changeReason` is deprecated and will be removed in version `3.0.0`
 
 2.9.0 (2020-04-23)
 ------------------